requirements.txt
# --- API & Server ---
fastapi==0.115.2
uvicorn==0.30.6
requests==2.32.3

# --- Data Handling & Machine Learning ---
pandas==2.2.3
numpy>=2.0.0
scikit-learn==1.5.2
joblib==1.4.2
tqdm==4.66.5

# --- Database (MySQL + ORM Layer) ---
sqlalchemy>=2.0.0
pymysql>=1.1.0

# --- Authentication & Security ---
passlib[bcrypt]==1.7.4

# --- Environment Configuration ---
python-dotenv>=1.0.1
